requirements.txt
RPi.GPIO
paho-mqtt~=1.6.1
smbus~=1.1.post2
Unidecode~=1.3.7